tests/wpt/tests/content-security-policy/resource-hints/prefetch-ignores-prefetch-src.sub.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='/common/utils.js'></script>
<script src='/content-security-policy/support/testharness-helper.js'></script>
<script src="/content-security-policy/support/prefetch-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="default-src 'none'; script-src 'self' 'unsafe-inline'; prefetch-src http://{{domains[www2]}}:{{ports[http][0]}}"/>

<script>
  promise_test(async (t) => {
    assert_false(await
      try_to_prefetch('http://{{domains[www2]}}:{{ports[http][0]}}/common/dummy.xml',
      t));
  }, "Prefetch should fail when restricted by default-src and allowed by " +
     "unsupported prefetch-src directive (prefetch-src should be ignored)");
</script>
</head>
<body></body>
</html>
