LayoutTests/imported/w3c/web-platform-tests/content-security-policy/resource-hints/prefetch-ignores-prefetch-src.sub-expected.txt

CONSOLE MESSAGE: Blocked access to external URL http://www2.localhost:8800/common/dummy.xml?pipe=%7Cheader%28Cache-Control%2C+max-age%3D604800%29%7Cheader%28Access-Control-Allow-Origin%2C+*%29%7Cheader%28Timing-Allow-Origin%2C+*%29&uuid=13eb39a2-c31f-43e4-bbdc-cd1bba517e0c

Harness Error (TIMEOUT), message = null

TIMEOUT Prefetch should fail when restricted by default-src and allowed by unsupported prefetch-src directive (prefetch-src should be ignored) Test timed out
